simpleBackend/requirements.txt
django == 1.9.6
djangorestframework == 3.3.3

